Add a test.out file
The current testing strategy is stupid, it either requires you do have a
failing mvn validate to test with or to cat the test.out to cargo run:
`cat test.out | cargo run`.

This should be worked on but until that investment is made, add a test
file to make it easier for people to run the test.