Always pick the highest version of a dependency
Before this change the version picker and parser machinery could miss
higher versions in some edge cases. One such edge case was a managed
dependency that had a pinned version but a dependency required a higher
version.

In this commit we remove the requirement to have non whitespace
characters present before the coordinates. This enables us to capture
the case for the managed dependency edge case described above. We also
make sure to escape the string that we build the regex from.

The last chagne to the paring machinery is nopt just picking the frist
non full match of the regex, but all submatches for the version regex,
while the old logic worked in most cases, the relaxing of non white
space characters combined with walking through all submatched is what
enables us to support the manged dependency edge case descibed above.